+++ CodexLib/Classes/AnyRef.cs
@@ -124,12 +124,14 @@ namespace CodexLib
 
         public static AnyRef ToAny(object obj)
         {
+            if (obj is AnyRef any)
+                return any;
             if (obj is string str)
-                return new AnyRef() { deserializedGuid = BlueprintGuid.Parse(str) };
-            else if (obj is BlueprintReferenceBase bp)
-                return new AnyRef() { deserializedGuid = bp.deserializedGuid };
-            else if (obj is SimpleBlueprint sb)
-                return new AnyRef() { deserializedGuid = sb.AssetGuid };
+                return new AnyRef { deserializedGuid = BlueprintGuid.Parse(str) };
+            if (obj is BlueprintReferenceBase bp)
+                return new AnyRef { deserializedGuid = bp.deserializedGuid, Cached = bp.Cached };
+            if (obj is SimpleBlueprint sb)
+                return new AnyRef { deserializedGuid = sb.AssetGuid, Cached = sb };
             Helper.PrintError($"AnyRef could not resolve type '{obj?.GetType()}'");
             return null;
         }